move active learning functions into backend.extract_frames module
